Update and rename PyData/pyData.py to pyData.py
